change for-each by java8, added work version delete method and refactoring code

+++ logic/src/main/java/com/epam/esm/repository/api/BaseRepository.java
@@ -14,5 +14,5 @@ public interface BaseRepository<T> {
 
     T update(T entity);
 
-    void deleteById(long id);
+    void deleteById(int id);
 }
